mejore el diseño

+++ PlataformaVideojuegos/src/main/java/org/example/model/dto/purchase/PurchaseDTO.java
@@ -2,7 +2,6 @@ package org.example.model.dto.purchase;
 
 import org.example.model.dto.game.GameDTO;
 import org.example.model.dto.user.UserDTO;
-import org.example.model.paymentMethods.IPaymentMethod;
 
 import java.time.LocalDate;
 
@@ -15,7 +14,7 @@ public record PurchaseDTO(
         LocalDate purchaseDate,
         IPaymentMethod paymentMethod,
         float priceWithoutDiscount,
-        int discountApplicated,
+        float discountApplicated,
         PurchaseState satate) {
 
 }
